requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

discord.py==2.4.0

SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2    # optional, not used by default

a2s==1.5.0         # Steam A2S (UDP) query
rcon==2.4.7        # Source RCON client

python-dotenv==1.0.1
paramiko==3.2.0    # SFTP client
aiohttp==3.9.4     # HTTP client for async operations
httpx==0.27.0      # HTTP client for async operations
pydantic==2.7.3    # Data validation and settings management
redis==4.6.0       # Redis client for caching
